Add tests
Adds a lot of BATS tests for the Buildkite Plugin tester, as well as
some unit tests for the Bash scripts. Additionally, adds more
affirmative testing of both the environment and pre-exit hooks in the
context of an *actual* invocation of the plugin in Buildkite.

Also modifies some of the code to make it more testable.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>