Allow authentication role name to be explicitly set
Previously, we relied solely on internal logic to map a Buildkite
queue name to a Vault authentication role name. However, due to some
changes in Vault that have made it necessary to modify how we model
things in Vault, this logic will need to change.

In order to ease this transition, we add the ability to concretely
specify the name of the authentication role, overriding this internal
logic.

This helps to further generalize this plugin, and provides an "escape
hatch" for the future, should we ever need it.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>