Add an `always-pull` option
We currently define `hashicorp/vault:latest` to be the default image
to use for access to a `vault` binary. However, this means that the
first time an agent runs this plugin, whatever the current value of
`latest` is will be the image used for the life of that agent. This
isn't an issue for short-lived agents, but could be a concern with
longer-lived agents.

Here, we add an `always-pull` option to the plugin, which allows users
to have additional control over how the image is resolved. If enabled,
an explicit `docker pull` is performed prior to any `docker run`
operations. This enables users to get the _true_ `latest` image,
regardless of what may have been downloaded by the agent at any point
in the past. It defaults to `false`, and is thus backwards compatible.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>