Output the Vault image ID and version for visibility / debugging
Rather than simply printing out the name of the Vault image being
used, we now also print the ID of the image, as well as the output of
`vault --version`.

Since the `latest` tag will change over time, it can be good to know
more concretely which image and version of Vault is being used.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>